// src/nodecg/generated/spreadsheetStatus.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SpreadsheetStatus {
	enabled: boolean;
	lastSynced?: string;
	lastError?: string;
}